practice_1/project/requirements.txt
amqp==2.5.2
billiard==3.6.1.0
celery==4.3.0
Django==2.1.5
django-autoslug==1.9.6
django-filter==2.2.0
djangorestframework==3.11.0
importlib-metadata==1.1.0
kombu==4.6.6
more-itertools==8.0.0
Pillow==6.2.1
pytz==2019.3
sqlparse==0.3.0
vine==1.3.0
zipp==0.6.0
